+++ retroauth-android/src/main/java/com/andretietz/retroauth/AuthenticationActivity.java
@@ -80,11 +80,12 @@ public abstract class AuthenticationActivity extends AppCompatActivity {
         }
         accountType = intent.getStringExtra(AccountManager.KEY_ACCOUNT_TYPE);
         if (accountType == null)
-            throw new RuntimeException(
+            throw new IllegalStateException(
                     String.format(
-                            "This Activity cannot be started without the \"%s\" extra in the intent!",
-                            AccountManager.KEY_ACCOUNT_TYPE));
-        accountName = intent.getStringExtra(AccountManager.KEY_ACCOUNT_NAME);
+                            "This Activity cannot be started without the \"%s\" extra in the intent! "
+                                    + "Use the \"addAccount\"-Method of the \"%s\" for opening the Login manually.",
+                            AccountManager.KEY_ACCOUNT_TYPE, AuthAccountManager.class.getSimpleName()));
+        //accountName = intent.getStringExtra(AccountManager.KEY_ACCOUNT_NAME);
         tokenType = intent.getStringExtra(AccountAuthenticator.KEY_TOKEN_TYPE);
         accountManager = AccountManager.get(this);
 
@@ -148,9 +149,7 @@ public abstract class AuthenticationActivity extends AppCompatActivity {
         resultBundle.putString(AccountManager.KEY_ACCOUNT_NAME, account.name);
         SharedPreferences preferences = getSharedPreferences(accountType, Context.MODE_PRIVATE);
         preferences.edit().putString(AuthAccountManager.RETROAUTH_ACCOUNTNAME_KEY, account.name).apply();
-        if (finishActivity) {
-            finish();
-        }
+        if (finishActivity) finish();
     }
 
     /**
@@ -173,7 +172,7 @@ public abstract class AuthenticationActivity extends AppCompatActivity {
      * @return The account if found, or a newly created one
      */
     @NonNull
-    @SuppressWarnings("unused")
+    @SuppressWarnings({"unused", "MissingPermission"})
     protected Account createOrGetAccount(@NonNull String accountName) {
         // if this is a relogin
         Account[] accountList = accountManager.getAccountsByType(accountType);
@@ -192,6 +191,7 @@ public abstract class AuthenticationActivity extends AppCompatActivity {
      *
      * @param account to remove
      */
+    @SuppressWarnings("deprecation")
     protected void removeAccount(@NonNull Account account) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP_MR1) {
             accountManager.removeAccount(account, null, null, null);
@@ -233,6 +233,7 @@ public abstract class AuthenticationActivity extends AppCompatActivity {
      * @return account name of the user
      */
     @Nullable
+    @Deprecated
     @SuppressWarnings("unused")
     protected String getAccountName() {
         return accountName;